Constructor of the featscreen class

This is the constructor of the featscreen class.

Usage

featscreen(
  method = c("cor.test", "t.test", "w.test", "anova", "kruskal.wallis", "chisq.test",
    "coxph", "moderated.t", "moderated.F", "sam.test", "missing.value", "above.median",
    "above.minimum", "median", "variability"),
  multi = character(),
  selection = c("cutoff", "rank", "percentile", "fpr", "fdr"),
  summary = character(),
  n = integer(),
  features = character(),
  keep = vector(),
  ranks = integer()
)

Arguments

method

character string, the id of the used screening method.

multi

(optional) character string, the id of the used multi-response aggregation method.

selection

character string, the id of the used selection method.

summary

character string, a textual summary of the screening.

n

integer, the dimension of the feature space.

features

string vector, the names of the features constituting the feature space.

keep

integer or logical vector, the features to keep.

ranks

(optional) integer vector, the ordering indices.

Value

An object of class featscreen.

Details

An object of class featscreen is a named list with eight elements:

method

the id of the used screening method

multi

the id of the used multi-response aggregation method

selection

the id of the used selection method

summary

a textual summary of the screening

n

the dimension of the feature space

features

the feature names

keep

the features to keep

ranks

the feature ranks

Functions to facilitate access to the data stored in a resampling object are available:

?getScreeningMethodId: returns the screening method id

?getMultiresponseAggregationMethodId: returns the multi-response aggregation method id

?getSelectionMethodId: returns the selection method id

?getSummary: returns the textual summary of the screening

?getFeatureDimensionality: returns the dimension of the feature space

?getFeatureNames: returns the feature names

?getScreenedFeatures: returns the features to keep

?getFeatureRanks: returns the feature ranks

Other useful functions include:

print: print a summary of the featscreen object

Author

Alessandro Barberis

Examples

#default
featscreen()

#featscreen object
featscreen(
 method = 'cor.test',
 selection = 'cutoff',
 summary = '3 out of 5 features selected by a cutoff.',
 n = 5,
 keep = c(TRUE,FALSE,TRUE,FALSE,TRUE),
 features = paste0("f", seq_len(5)),
 ranks = c(1,5,3,4,2)
)
#> 
#> 3 out of 5 features selected by a cutoff.
#> 
#> Top 5 ranked features: f1, f5, f3, f4, f2
#>