Available Screening Methods

This function returns the currently available screening methods.

Usage

listAvailableScreeningMethods(
  x = c("all", "unsupervised", "supervised"),
  show.wrapper = FALSE,
  show.data.type = FALSE,
  simplify = FALSE,
  show.function = FALSE,
  show.category = FALSE
)

Arguments

x

character string indicating whether to return all screening methods (all), the unsupervised screening methods (unsupervised) or the supervised screening methods (supervised).

show.wrapper

logical indicating whether to return the ids for the supervised screening functions with or without wrapper functions.

show.data.type

logical indicating whether to show the supported data types.

simplify

logical indicating whether to format the data for readability.

show.function

logical indicating whether to show the function name.

show.category

logical indicating whether to show the internally given category.

Value

A data frame with at least two columns:

id

the id of the screening method, to be used in the function calls

name

the name of the screening method

Author

Alessandro Barberis

Examples

#List available methods
listAvailableScreeningMethods()
#>                id
#> 1   missing.value
#> 2    above.median
#> 3   above.minimum
#> 4          median
#> 5     variability
#> 6         pearson
#> 7        spearman
#> 8         kendall
#> 9    t.test.equal
#> 10 t.test.unequal
#> 11  t.test.paired
#> 12 w.test.ranksum
#> 13  w.test.paired
#> 14    anova.equal
#> 15  anova.unequal
#> 16 kruskal.wallis
#> 17     chisq.test
#> 18          coxph
#> 19    moderated.t
#> 20    moderated.F
#> 21       sam.test
#>                                                                       name
#> 1                                                      missing value ratio
#> 2                                             above-median frequency ratio
#> 3                                            above-minimum frequency ratio
#> 4                                                             median value
#> 5                                                              variability
#> 6                  Pearson's product moment correlation coefficient t-test
#> 7                           Spearman's rank correlation coefficient t-test
#> 8                            Kendall's rank correlation coefficient t-test
#> 9                                       two-sample Student's pooled t-test
#> 10 two-sample t-test with the Welch modification to the degrees of freedom
#> 11                                      paired two-sample Student's t-test
#> 12                                          two-sample Mann-Whitney U-test
#> 13                             paired two-sample Wilcoxon signed-rank test
#> 14                                     one-way analysis of variance F-test
#> 15               one-way analysis of variance F-test with Welch correction
#> 16                                                   Kruskal-Wallis H-test
#> 17                                                       Pearson's χ²-test
#> 18                                    Cox PH regression coefficient z-test
#> 19                                        empirical Bayes moderated t-test
#> 20                                        empirical Bayes moderated F-test
#> 21                    significant analysis of microarrays permutation test

#List available methods and supported data types
listAvailableScreeningMethods(show.data.type = TRUE, simplify = TRUE)
#>                id
#> 1    above.median
#> 2   above.minimum
#> 3     anova.equal
#> 4   anova.unequal
#> 5      chisq.test
#> 6           coxph
#> 7         kendall
#> 8  kruskal.wallis
#> 9          median
#> 10  missing.value
#> 11    moderated.F
#> 12    moderated.t
#> 13        pearson
#> 14       sam.test
#> 15       spearman
#> 16   t.test.equal
#> 17  t.test.paired
#> 18 t.test.unequal
#> 19    variability
#> 20  w.test.paired
#> 21 w.test.ranksum
#>                                                                       name
#> 1                                             above-median frequency ratio
#> 2                                            above-minimum frequency ratio
#> 3                                      one-way analysis of variance F-test
#> 4                one-way analysis of variance F-test with Welch correction
#> 5                                                        Pearson's χ²-test
#> 6                                     Cox PH regression coefficient z-test
#> 7                            Kendall's rank correlation coefficient t-test
#> 8                                                    Kruskal-Wallis H-test
#> 9                                                             median value
#> 10                                                     missing value ratio
#> 11                                        empirical Bayes moderated F-test
#> 12                                        empirical Bayes moderated t-test
#> 13                 Pearson's product moment correlation coefficient t-test
#> 14                    significant analysis of microarrays permutation test
#> 15                          Spearman's rank correlation coefficient t-test
#> 16                                      two-sample Student's pooled t-test
#> 17                                      paired two-sample Student's t-test
#> 18 two-sample t-test with the Welch modification to the degrees of freedom
#> 19                                                             variability
#> 20                             paired two-sample Wilcoxon signed-rank test
#> 21                                          two-sample Mann-Whitney U-test
#>    numerical categorical multiresponse survival
#> 1          x                                   
#> 2          x                                   
#> 3                      x             x         
#> 4                      x             x         
#> 5                      x             x         
#> 6                                             x
#> 7          x                                   
#> 8                      x             x         
#> 9          x                                   
#> 10         x           x             x        x
#> 11         x           x             x         
#> 12         x           x             x         
#> 13         x                                   
#> 14         x           x             x        x
#> 15         x                                   
#> 16                     x                       
#> 17                     x                       
#> 18                     x                       
#> 19         x                                   
#> 20                     x                       
#> 21                     x                       

#List available methods and function names
listAvailableScreeningMethods(show.function = TRUE)
#>                id
#> 1    above.median
#> 2   above.minimum
#> 3     anova.equal
#> 4   anova.unequal
#> 5      chisq.test
#> 6           coxph
#> 7         kendall
#> 8  kruskal.wallis
#> 9          median
#> 10  missing.value
#> 11    moderated.F
#> 12    moderated.t
#> 13        pearson
#> 14       sam.test
#> 15       spearman
#> 16   t.test.equal
#> 17  t.test.paired
#> 18 t.test.unequal
#> 19    variability
#> 20  w.test.paired
#> 21 w.test.ranksum
#>                                                                       name
#> 1                                             above-median frequency ratio
#> 2                                            above-minimum frequency ratio
#> 3                                      one-way analysis of variance F-test
#> 4                one-way analysis of variance F-test with Welch correction
#> 5                                                        Pearson's χ²-test
#> 6                                     Cox PH regression coefficient z-test
#> 7                            Kendall's rank correlation coefficient t-test
#> 8                                                    Kruskal-Wallis H-test
#> 9                                                             median value
#> 10                                                     missing value ratio
#> 11                                        empirical Bayes moderated F-test
#> 12                                        empirical Bayes moderated t-test
#> 13                 Pearson's product moment correlation coefficient t-test
#> 14                    significant analysis of microarrays permutation test
#> 15                          Spearman's rank correlation coefficient t-test
#> 16                                      two-sample Student's pooled t-test
#> 17                                      paired two-sample Student's t-test
#> 18 two-sample t-test with the Welch modification to the degrees of freedom
#> 19                                                             variability
#> 20                             paired two-sample Wilcoxon signed-rank test
#> 21                                          two-sample Mann-Whitney U-test
#>                    function
#> 1   rowAboveMedianFreqRatio
#> 2      rowAboveMinFreqRatio
#> 3    rowEqualVarOneWayAnova
#> 4  rowUnequalVarOneWayAnova
#> 5           rowPearsonChiSq
#> 6                  rowCoxPH
#> 7             rowKendallCor
#> 8          rowKruskalWallis
#> 9                rowMedians
#> 10     rowMissingValueRatio
#> 11  rowModeratedOneWayAnova
#> 12            rowModeratedT
#> 13            rowPearsonCor
#> 14         rowSamStatistics
#> 15           rowSpearmanCor
#> 16             rowEqualVarT
#> 17               rowPairedT
#> 18           rowUnequalVarT
#> 19           rowVariability
#> 20       rowPairedWilcoxonT
#> 21             rowWilcoxonT