Fit Univariate Proportional Hazards Regression Model

Fits a Cox proportional hazards regression model for each row vector in x.

Usage

rowCoxPH(x, y, ...)

Arguments

x

independent variable

y

data.frame with two columns, time and status

...

further arguments to coxph

Value

A list containing two elements:

statistic

A numeric vector, the values of the test statistic

significance

A numeric vector, the p-values of the selected test

Details

It is a wrapper to coxph function.

Author

Alessandro Barberis

Examples

#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 7), 100, 7)
y = data.frame(
  time = c(4,3,1,1,2,2,3),
  status = c(1,1,1,0,1,1,0)
)

#Compute
rowCoxPH(x = x, y = y)
#> $statistic
#>   [1] 3.221007e+00 2.935206e+00 1.140501e+00 4.454102e-01 1.190286e+00
#>   [6] 8.916962e-01 8.049666e-01 1.590121e+00 7.349304e-01 3.560573e+00
#>  [11] 4.310211e+00 1.771256e+00 7.358665e-01 7.834584e-01 5.595023e-01
#>  [16] 8.517362e-01 1.502229e+00 1.074111e+00 1.020083e+00 9.756255e-01
#>  [21] 4.227505e-01 1.608548e+01 4.521620e+00 2.692001e-01 1.074576e+00
#>  [26] 3.624716e-01 1.086745e+00 9.113874e-01 1.697756e+01 2.533270e+00
#>  [31] 1.275738e+00 3.907546e-01 4.542239e-01 1.779742e-01 9.083797e-01
#>  [36] 6.280007e-01 1.396546e+00 6.101014e+00 8.530357e-01 2.840826e+00
#>  [41] 1.215659e+00 1.094521e-06 2.701355e-01 1.082217e+00 1.052698e+00
#>  [46] 7.733836e-01 1.457949e+00 1.568748e+00 1.024923e+00 1.813662e+00
#>  [51] 8.365681e-01 9.254171e-01 1.622466e+00 8.343404e-01 8.929156e-01
#>  [56] 8.463514e-01 1.002840e+00 8.800736e-01 4.623974e+00 2.437298e-01
#>  [61] 2.909040e-01 1.164090e+00 1.735525e+00 9.390681e-01 6.240458e-01
#>  [66] 1.230961e+00 7.307977e-01 8.905954e-01 1.734010e+00 8.942129e-01
#>  [71] 1.189308e+00 4.097652e-01 1.915732e+00 8.902901e-01 1.525230e+00
#>  [76] 1.242867e+00 3.009826e+00 1.016969e+00 3.517028e-01 4.517073e-01
#>  [81] 1.057733e-01 1.384930e+00 3.401540e-01 6.624688e-01 6.530762e-01
#>  [86] 1.633748e+00 1.378153e+00 1.396723e+00 7.800584e-01 1.153907e+00
#>  [91] 1.355294e+00 4.010554e-01 2.730155e+00 1.650111e+01 1.210385e+00
#>  [96] 9.941296e+00 9.080608e-01 4.625763e-02 1.722859e-01 8.008059e-01
#> 
#> $significance
#>   [1] 0.35284158 0.23216275 0.73758050 0.30118757 0.85494752 0.79780513
#>   [7] 0.57710663 0.26094473 0.69492137 0.31890517 0.26710108 0.66551492
#>  [13] 0.77760548 0.61193232 0.73318231 0.80078024 0.54018698 0.92859258
#>  [19] 0.97586279 0.94423064 0.26460129 0.10786105 0.17282449 0.11431238
#>  [25] 0.89624827 0.08181236 0.75986679 0.89146060 0.21163423 0.42229480
#>  [31] 0.70298142 0.27837110 0.33107439 0.07258021 0.85710750 0.58156520
#>  [37] 0.66554431 0.10609623 0.76369326 0.34983203 0.70231073 0.16166355
#>  [43] 0.22807359 0.90435796 0.91378796 0.72984889 0.56438354 0.46137108
#>  [49] 0.98487894 0.39205220 0.68143883 0.87972285 0.24956478 0.73447671
#>  [55] 0.88573483 0.82775947 0.99658780 0.82385224 0.17635549 0.05941668
#>  [61] 0.09367397 0.84519183 0.40927596 0.93408775 0.14305795 0.63458694
#>  [67] 0.59578763 0.86391024 0.30615053 0.83262541 0.74192289 0.45133100
#>  [73] 0.41624233 0.77796538 0.56882428 0.63632036 0.24414165 0.98456609
#>  [79] 0.29858647 0.21476206 0.08861988 0.56581889 0.06272944 0.36847209
#>  [85] 0.32801611 0.25534905 0.62283278 0.52537683 0.62102484 0.61296241
#>  [91] 0.49592007 0.17068410 0.22844039 0.13981152 0.64839154 0.16112832
#>  [97] 0.86323414 0.15522274 0.07572453 0.63635069
#>