Fit an Analysis of Variance Model

Computes the one-way analysis of variance test for each feature.

See the Details section below for further information.

Usage

rowEqualVarOneWayAnova(x, g)

Arguments

x

numeric matrix.

g

a vector specifying group membership for each observation of x.

Value

A list containing two elements:

statistic

A numeric vector, the values of the test statistic

significance

A numeric vector, the p-values of the selected test

Details

It is a wrapper to row_oneway_equalvar function.

Author

Alessandro Barberis

Examples

#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 20), 100, 20)
g = sample(c(0,1), 20, replace = TRUE)

#Compute
rowEqualVarOneWayAnova(x = x, g = g)
#> $statistic
#>   [1] 6.175662e-01 1.142795e+00 3.373973e+00 7.497984e-02 1.391426e+00
#>   [6] 3.693834e-01 2.206839e-03 1.337518e-01 1.393143e-01 1.867596e-04
#>  [11] 1.052904e+00 1.307385e+01 5.739056e-02 7.198392e-01 8.443845e+00
#>  [16] 4.173536e-02 3.927375e-02 8.502934e+00 3.000569e-01 1.298206e+00
#>  [21] 7.850476e-01 1.219352e-03 2.220413e-01 1.085044e-01 2.457722e+00
#>  [26] 2.851262e+00 5.001434e-01 3.050954e+00 6.751124e-01 1.425549e+00
#>  [31] 9.876122e-02 2.123565e-01 1.786957e+00 5.387085e-02 1.830630e-01
#>  [36] 7.907822e+00 1.404200e-01 1.887882e+00 9.987513e-01 1.153725e+00
#>  [41] 7.991206e-02 1.595998e+00 2.788211e-01 1.637761e+00 6.795809e+00
#>  [46] 6.509314e-02 7.023671e-02 2.777753e-02 1.698515e+00 1.805773e+00
#>  [51] 1.657147e+00 1.561988e-03 9.443359e-01 3.021035e-01 3.829611e-01
#>  [56] 1.801679e+00 9.344037e-01 4.171424e-01 1.159628e+00 1.463029e-01
#>  [61] 1.619973e-03 7.273319e-02 5.191157e-02 5.580119e-03 1.368136e-01
#>  [66] 3.483534e-01 3.600077e-01 2.543502e-01 1.289645e+00 2.090509e-03
#>  [71] 3.617872e+00 3.677760e+00 6.072455e-01 7.851055e-01 2.769109e+00
#>  [76] 2.926726e-01 2.105485e+00 1.654725e+00 3.870242e-02 6.183033e-01
#>  [81] 1.645710e+00 9.626418e+00 1.629068e+00 8.806703e-02 8.105337e-02
#>  [86] 3.992556e-01 2.373407e-01 1.654598e-01 1.068649e+00 5.158298e-03
#>  [91] 2.883852e+00 7.927073e-01 1.372184e+00 5.624146e-02 1.004130e-01
#>  [96] 5.893066e-02 2.731018e-01 1.656353e-01 1.971405e+00 1.539184e-01
#> 
#> $significance
#>   [1] 0.442175859 0.299189535 0.082803822 0.787336975 0.253520606 0.550932350
#>   [7] 0.963048844 0.718835880 0.713323057 0.989246781 0.318432101 0.001976585
#>  [13] 0.813375570 0.407338394 0.009425943 0.840417168 0.845128304 0.009221409
#>  [19] 0.590575494 0.269477549 0.387286174 0.972528477 0.643149627 0.745655395
#>  [25] 0.134359728 0.108550730 0.488492833 0.097731175 0.422025211 0.247990429
#>  [31] 0.756934565 0.650445702 0.197948760 0.819077950 0.673829370 0.011533983
#>  [37] 0.712242269 0.186310456 0.330859053 0.296959476 0.780641688 0.222590206
#>  [43] 0.603924431 0.216882408 0.017839470 0.801512307 0.794003581 0.869490854
#>  [49] 0.208907405 0.195711365 0.214296244 0.968909274 0.344043779 0.589321984
#>  [55] 0.543775817 0.196195449 0.346530494 0.526515332 0.295764322 0.706572649
#>  [61] 0.968337774 0.790465508 0.822337813 0.941277182 0.715785572 0.562383956
#>  [67] 0.555980978 0.620146075 0.271008680 0.964035200 0.073287907 0.071150524
#>  [73] 0.445951945 0.387269020 0.113411033 0.595145435 0.163976809 0.214617187
#>  [79] 0.846243578 0.441908164 0.215817215 0.006143642 0.218054907 0.770046205
#>  [85] 0.779124524 0.535417486 0.632013346 0.688974669 0.314940928 0.943536143
#>  [91] 0.106691505 0.385028695 0.256710368 0.815216383 0.754980309 0.810938795
#>  [97] 0.607631515 0.688819158 0.177317919 0.699425299
#>