Fit an Analysis of Variance Model

Computes the one-way analysis of variance test for each feature.

See the Details section below for further information.

Usage

rowOneWayAnova(x, g, var = c("unequal", "equal"))

Arguments

x

numeric matrix.

g

a vector specifying group membership for each observation of x.

var

character string, indicating whether the groups have "equal" or "unequal" (default) variances

Value

A list containing two elements:

statistic

A numeric vector, the values of the test statistic

significance

A numeric vector, the p-values of the selected test

Details

It is a wrapper to row_oneway_equalvar function.

Author

Alessandro Barberis

Examples

#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 20), 100, 20)
g = sample(c(0,1), 20, replace = TRUE)

#Compute one-way analysis of variance
rowOneWayAnova(x = x, g = g, var = "equal")
#> $statistic
#>   [1] 6.175662e-01 1.142795e+00 3.373973e+00 7.497984e-02 1.391426e+00
#>   [6] 3.693834e-01 2.206839e-03 1.337518e-01 1.393143e-01 1.867596e-04
#>  [11] 1.052904e+00 1.307385e+01 5.739056e-02 7.198392e-01 8.443845e+00
#>  [16] 4.173536e-02 3.927375e-02 8.502934e+00 3.000569e-01 1.298206e+00
#>  [21] 7.850476e-01 1.219352e-03 2.220413e-01 1.085044e-01 2.457722e+00
#>  [26] 2.851262e+00 5.001434e-01 3.050954e+00 6.751124e-01 1.425549e+00
#>  [31] 9.876122e-02 2.123565e-01 1.786957e+00 5.387085e-02 1.830630e-01
#>  [36] 7.907822e+00 1.404200e-01 1.887882e+00 9.987513e-01 1.153725e+00
#>  [41] 7.991206e-02 1.595998e+00 2.788211e-01 1.637761e+00 6.795809e+00
#>  [46] 6.509314e-02 7.023671e-02 2.777753e-02 1.698515e+00 1.805773e+00
#>  [51] 1.657147e+00 1.561988e-03 9.443359e-01 3.021035e-01 3.829611e-01
#>  [56] 1.801679e+00 9.344037e-01 4.171424e-01 1.159628e+00 1.463029e-01
#>  [61] 1.619973e-03 7.273319e-02 5.191157e-02 5.580119e-03 1.368136e-01
#>  [66] 3.483534e-01 3.600077e-01 2.543502e-01 1.289645e+00 2.090509e-03
#>  [71] 3.617872e+00 3.677760e+00 6.072455e-01 7.851055e-01 2.769109e+00
#>  [76] 2.926726e-01 2.105485e+00 1.654725e+00 3.870242e-02 6.183033e-01
#>  [81] 1.645710e+00 9.626418e+00 1.629068e+00 8.806703e-02 8.105337e-02
#>  [86] 3.992556e-01 2.373407e-01 1.654598e-01 1.068649e+00 5.158298e-03
#>  [91] 2.883852e+00 7.927073e-01 1.372184e+00 5.624146e-02 1.004130e-01
#>  [96] 5.893066e-02 2.731018e-01 1.656353e-01 1.971405e+00 1.539184e-01
#> 
#> $significance
#>   [1] 0.442175859 0.299189535 0.082803822 0.787336975 0.253520606 0.550932350
#>   [7] 0.963048844 0.718835880 0.713323057 0.989246781 0.318432101 0.001976585
#>  [13] 0.813375570 0.407338394 0.009425943 0.840417168 0.845128304 0.009221409
#>  [19] 0.590575494 0.269477549 0.387286174 0.972528477 0.643149627 0.745655395
#>  [25] 0.134359728 0.108550730 0.488492833 0.097731175 0.422025211 0.247990429
#>  [31] 0.756934565 0.650445702 0.197948760 0.819077950 0.673829370 0.011533983
#>  [37] 0.712242269 0.186310456 0.330859053 0.296959476 0.780641688 0.222590206
#>  [43] 0.603924431 0.216882408 0.017839470 0.801512307 0.794003581 0.869490854
#>  [49] 0.208907405 0.195711365 0.214296244 0.968909274 0.344043779 0.589321984
#>  [55] 0.543775817 0.196195449 0.346530494 0.526515332 0.295764322 0.706572649
#>  [61] 0.968337774 0.790465508 0.822337813 0.941277182 0.715785572 0.562383956
#>  [67] 0.555980978 0.620146075 0.271008680 0.964035200 0.073287907 0.071150524
#>  [73] 0.445951945 0.387269020 0.113411033 0.595145435 0.163976809 0.214617187
#>  [79] 0.846243578 0.441908164 0.215817215 0.006143642 0.218054907 0.770046205
#>  [85] 0.779124524 0.535417486 0.632013346 0.688974669 0.314940928 0.943536143
#>  [91] 0.106691505 0.385028695 0.256710368 0.815216383 0.754980309 0.810938795
#>  [97] 0.607631515 0.688819158 0.177317919 0.699425299
#> 

#Compute one-way analysis of variance with Welch correction
rowOneWayAnova(x = x, g = g)
#> $statistic
#>   [1] 8.916031e-01 1.099033e+00 2.582175e+00 7.518515e-02 2.035640e+00
#>   [6] 2.677994e-01 1.894486e-03 1.169271e-01 1.702897e-01 2.012154e-04
#>  [11] 1.646617e+00 1.381326e+01 6.321682e-02 9.411447e-01 7.969342e+00
#>  [16] 4.603139e-02 3.435455e-02 6.960313e+00 3.138337e-01 9.638504e-01
#>  [21] 7.114030e-01 1.306512e-03 1.621167e-01 1.049471e-01 2.100979e+00
#>  [26] 2.268712e+00 4.226855e-01 3.435337e+00 9.029074e-01 1.084025e+00
#>  [31] 1.078762e-01 2.141863e-01 2.344181e+00 4.181963e-02 1.954188e-01
#>  [36] 7.172773e+00 2.433598e-01 1.620515e+00 8.769713e-01 1.722764e+00
#>  [41] 6.437042e-02 2.424702e+00 3.825107e-01 1.301477e+00 5.048114e+00
#>  [46] 6.496613e-02 5.148398e-02 4.029883e-02 2.587207e+00 1.297710e+00
#>  [51] 1.575327e+00 1.512358e-03 9.406518e-01 3.464681e-01 3.096683e-01
#>  [56] 1.869932e+00 1.037205e+00 4.234893e-01 1.097537e+00 1.137047e-01
#>  [61] 1.489393e-03 7.355277e-02 6.447568e-02 4.547635e-03 1.660285e-01
#>  [66] 3.691790e-01 4.379864e-01 2.588221e-01 1.346407e+00 2.446432e-03
#>  [71] 3.355627e+00 5.909533e+00 6.742883e-01 6.268341e-01 3.644805e+00
#>  [76] 2.634236e-01 2.153229e+00 1.765084e+00 3.484087e-02 5.815123e-01
#>  [81] 1.451825e+00 9.365198e+00 1.074278e+00 7.870752e-02 7.401629e-02
#>  [86] 3.590715e-01 2.189886e-01 1.423922e-01 1.074156e+00 3.446317e-03
#>  [91] 2.228390e+00 6.356445e-01 1.840563e+00 5.215740e-02 1.057160e-01
#>  [96] 4.573399e-02 3.371221e-01 2.079691e-01 1.780421e+00 1.239888e-01
#> 
#> $significance
#>   [1] 0.357546542 0.315579577 0.143541088 0.788427889 0.170785159 0.618448064
#>   [7] 0.966133510 0.739258268 0.685262850 0.988885750 0.216259927 0.002463609
#>  [13] 0.805083386 0.345363101 0.016020246 0.833163326 0.856546887 0.025884019
#>  [19] 0.584709770 0.353520348 0.417237047 0.971688698 0.697406692 0.751624963
#>  [25] 0.177787064 0.165715579 0.530443544 0.084087040 0.354896341 0.325946862
#>  [31] 0.747422508 0.651425358 0.143801000 0.842570767 0.665393206 0.021734843
#>  [37] 0.628924895 0.231627831 0.370324866 0.205931909 0.805242506 0.137089835
#>  [43] 0.544094849 0.282956217 0.053142710 0.803000924 0.825977103 0.843148800
#>  [49] 0.125400567 0.287042115 0.234229224 0.969626195 0.350744477 0.564883009
#>  [55] 0.590972139 0.194535992 0.325333058 0.526826696 0.316358083 0.743782125
#>  [61] 0.969904597 0.790624441 0.802696381 0.947643363 0.689041506 0.553564477
#>  [67] 0.517446270 0.619644873 0.266493436 0.961188009 0.093758391 0.026476980
#>  [73] 0.424983925 0.448468792 0.072901699 0.618188949 0.166404162 0.205878620
#>  [79] 0.855414010 0.461230641 0.254846551 0.009952655 0.332471228 0.784446871
#>  [85] 0.790645951 0.561485060 0.648875791 0.713721076 0.319653299 0.954712246
#>  [91] 0.170304775 0.445296545 0.192036848 0.823479844 0.750119182 0.835487756
#>  [97] 0.569398045 0.654248100 0.209624649 0.732602584
#>