Fit an Analysis of Variance Model

Computes the one-way analysis of variance test with Welch correction for each feature.

See the Details section below for further information.

Usage

rowUnequalVarOneWayAnova(x, g)

Arguments

x

numeric matrix.

g

a vector specifying group membership for each observation of x.

Value

A list containing two elements:

statistic

A numeric vector, the values of the test statistic

significance

A numeric vector, the p-values of the selected test

Details

It is a wrapper to row_oneway_welch function.

Author

Alessandro Barberis

Examples

#Seed
set.seed(1010)

#Data
x = matrix(rnorm(100 * 30), 100, 30)
g = sample(c(0,1,3), 30, replace = TRUE)

#Compute
rowUnequalVarOneWayAnova(x = x, g = g)
#> $statistic
#>   [1] 0.27405158 3.89755787 0.15995295 0.04428048 1.21016157 3.13738429
#>   [7] 1.21132569 0.30314501 0.06536424 0.27034838 1.81854543 0.09974003
#>  [13] 0.94306636 0.25306108 0.11917790 3.49733094 0.31892021 0.64745671
#>  [19] 7.83017472 1.23149089 1.33290700 0.72626226 3.16646167 0.86935833
#>  [25] 0.32769916 1.22966668 0.63266639 0.57919345 0.55513369 1.13167953
#>  [31] 1.00383181 2.96319618 0.29120005 2.06766044 0.13624214 1.34812254
#>  [37] 1.35450883 1.65729529 0.27865841 1.02801038 0.52994833 0.96022558
#>  [43] 1.38845903 1.38640293 0.24486710 0.52216337 1.49795157 0.54106463
#>  [49] 0.30378567 2.84751364 0.30478612 0.60317983 2.12882223 0.47369717
#>  [55] 3.97067868 0.36705202 0.06088811 0.45967915 0.68580528 0.18963709
#>  [61] 0.23945692 2.65123546 0.24620342 0.36906925 6.56152951 4.32785960
#>  [67] 4.10174415 0.81806297 0.35585415 0.63503564 1.29572299 1.07245127
#>  [73] 0.83655390 0.90101120 0.60388897 0.88787726 0.48533291 0.13885067
#>  [79] 0.43276971 0.01290793 0.56822203 0.42039165 1.16871123 1.79548379
#>  [85] 0.05557961 1.14583940 1.38165665 1.75503872 0.40933621 0.50292595
#>  [91] 1.69329657 0.64663006 0.45673609 3.00077676 2.87089701 0.40512476
#>  [97] 0.49794967 0.55032820 0.64641643 0.56635595
#> 
#> $significance
#>   [1] 0.765128703 0.047722915 0.853794415 0.956853138 0.329712866 0.076720356
#>   [7] 0.331512766 0.744815388 0.937043445 0.768653582 0.194167642 0.905773515
#>  [13] 0.418840473 0.780352564 0.888850697 0.060274222 0.732186323 0.541627651
#>  [19] 0.004664803 0.322888503 0.300712397 0.503318477 0.075582972 0.445354809
#>  [25] 0.727352347 0.322161521 0.546065894 0.573487160 0.585264321 0.351710622
#>  [31] 0.392803244 0.083936660 0.752454768 0.164275661 0.873830459 0.294533395
#>  [37] 0.291174524 0.225156990 0.761589815 0.383819206 0.600598048 0.404953653
#>  [43] 0.288210595 0.287714027 0.785653523 0.606990532 0.256340471 0.597582418
#>  [49] 0.743519930 0.094778201 0.742080749 0.560505617 0.158571659 0.633230543
#>  [55] 0.048027809 0.700438331 0.941209960 0.642703741 0.521564098 0.829592345
#>  [61] 0.790637412 0.105154607 0.785658261 0.697571114 0.015286226 0.037990410
#>  [67] 0.039845277 0.471776077 0.708286531 0.547444428 0.305848259 0.367340056
#>  [73] 0.459273725 0.429560281 0.560665004 0.432236284 0.628445988 0.871866068
#>  [79] 0.658243511 0.987190253 0.581153373 0.665479629 0.343346405 0.201077604
#>  [85] 0.946158891 0.343571037 0.284362851 0.220449858 0.672801675 0.614531969
#>  [91] 0.216711252 0.540327826 0.644775271 0.083809447 0.087983607 0.676871533
#>  [97] 0.618947055 0.591570132 0.541766133 0.580991048
#>